Moved mapper outside from controller
